# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FortKnoxx Security Intelligence Platform
  description: 'Unified security scanning API: SAST, DAST, SCA, IaC, secrets, container, and AI-powered scanners (zero-day
    GNN, business logic, LLM adversarial, auth).'
  version: 1.0.0
paths:
...
  /api/repositories:
    get:
      tags:
      - repositories
      summary: List Repositories
      operationId: list_repositories_api_repositories_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Repository'
                type: array
                title: Response List Repositories Api Repositories Get
    post:
      summary: Add Repository
      description: Add a repository for scanning (supports both public repos and private repos with tokens)
      operationId: add_repository_api_repositories_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRepositoryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/repositories/{repo_id}:
    get:
      tags:
      - repositories
      summary: Get Repository
      operationId: get_repository_api_repositories__repo_id__get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - repositories
      summary: Delete Repository
      operationId: delete_repository_api_repositories__repo_id__delete
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete Repository Api Repositories  Repo Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/scans/{repo_id}:
    post:
      summary: Start Scan
      description: Start a security scan for a repository
      operationId: start_scan_api_scans__repo_id__post
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - scans
      summary: List Scans For Repo
      operationId: list_scans_for_repo_api_scans__repo_id__get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scan'
                title: Response List Scans For Repo Api Scans  Repo Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/scans/detail/{scan_id}:
    get:
      tags:
      - scans
      summary: Get Scan
      operationId: get_scan_api_scans_detail__scan_id__get
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scan'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/scans/{scan_id}:
    delete:
      tags:
      - scans
      summary: Delete Scan
      operationId: delete_scan_api_scans__scan_id__delete
      parameters:
      - name: scan_id
        in: path
        required: true
        schema:
          type: string
          title: Scan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Delete Scan Api Scans  Scan Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/reports/generate:
    post:
      tags:
      - reports
      summary: Generate Report
      operationId: generate_report_api_reports_generate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Generate Report Api Reports Generate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AddRepositoryRequest:
      properties:
        provider:
          $ref: '#/components/schemas/GitProvider'
        repo_url:
          type: string
          title: Repo Url
          description: Repository URL (e.g., https://github.com/owner/repo)
        auto_scan:
          type: boolean
          title: Auto Scan
          description: Enable auto-scan on push
          default: false
        branch:
          anyOf:
          - type: string
          - type: 'null'
          title: Branch
          description: Branch to scan (defaults to default branch)
        access_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
          description: Access token for private repos (optional for public repos)
        is_public:
          type: boolean
          title: Is Public
          description: Whether this is a public repository (no token required)
          default: false
      type: object
      required:
      - provider
      - repo_url
      title: AddRepositoryRequest
      description: Request model for adding a repository to scan
      example:
        auto_scan: true
        branch: main
        is_public: true
        provider: github
        repo_url: https://github.com/myorg/myrepo
...
    GitProvider:
      type: string
      enum:
      - github
      - gitlab
      title: GitProvider
      description: Supported Git providers
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ReportRequest:
      properties:
        repo_id:
          type: string
          title: Repo Id
        scan_id:
          type: string
          title: Scan Id
        format:
          type: string
          title: Format
          default: json
      type: object
      required:
      - repo_id
      - scan_id
      title: ReportRequest
    Repository:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        url:
          type: string
          title: Url
        access_token:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
        branch:
          type: string
          title: Branch
          default: main
        last_scan:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Scan
        scan_status:
          type: string
          title: Scan Status
          default: pending
        created_at:
          type: string
          format: date-time
          title: Created At
        security_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Security Score
        vulnerabilities_count:
          type: integer
          title: Vulnerabilities Count
          default: 0
        critical_count:
          type: integer
          title: Critical Count
          default: 0
        high_count:
          type: integer
          title: High Count
          default: 0
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
      type: object
      required:
      - name
      - url
      title: Repository
    Scan:
      properties:
        id:
          type: string
          title: Id
        repo_id:
          type: string
          title: Repo Id
        status:
          type: string
          title: Status
          default: pending
        started_at:
          type: string
          format: date-time
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        total_files:
          type: integer
          title: Total Files
          default: 0
        vulnerabilities_count:
          type: integer
          title: Vulnerabilities Count
          default: 0
        quality_issues_count:
          type: integer
          title: Quality Issues Count
          default: 0
        compliance_issues_count:
          type: integer
          title: Compliance Issues Count
          default: 0
        critical_count:
          type: integer
          title: Critical Count
          default: 0
        high_count:
          type: integer
          title: High Count
          default: 0
        medium_count:
          type: integer
          title: Medium Count
          default: 0
        low_count:
          type: integer
          title: Low Count
          default: 0
        security_score:
          type: integer
          title: Security Score
          default: 0
        quality_score:
          type: integer
          title: Quality Score
          default: 100
        compliance_score:
          type: integer
          title: Compliance Score
          default: 100
        scan_results:
          additionalProperties: true
          type: object
          title: Scan Results
      type: object
      required:
      - repo_id
      title: Scan
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...